+++ src/main/java/com/fordogs/order/presentation/response/OrderSearchSellerResponse.java
@@ -7,7 +7,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -28,7 +27,7 @@ public class OrderSearchSellerResponse {
     private OrderStatus orderStatus;
 
     @Schema(description = "주문 총 금액")
-    private BigDecimal orderTotalPrice;
+    private Integer orderTotalPrice;
 
     @Schema(description = "주문 날짜")
     private LocalDateTime orderDate;
